fix: isolate touch-Settings helpers out of shared SettingsList.h
applySettingToggle() and formatSettingValueText() were pure-additive
but still lived in SettingsList.h, an upstream-owned file (imported
QuranBook.h/UITheme.h it never needed otherwise). Moves both,
unchanged, into a new Midad-only MidadTouchSettingsSupport.h consumed
only by TouchSettingsGroupActivity.cpp, and restores SettingsList.h to
byte-identical with the PR base.
